revised models to match source changes

+++ models/marts/service_points.sql
@@ -50,4 +50,4 @@ left join {{ref('stg_service_point_labels')}} label_ty on label_ty.name = main.l
 left join {{ref('stg_service_point_labels')}} label_open on label_open.name = main.open_24_7
 left join {{ref('activity_details')}} label_ad on label_ad.id = main.id
 left join {{ref('operational_hours')}} label_op on label_op.id = main.id
-left join {{ ref('int_avg_service_rating') }} avg_rtg on main.id = avg_rtg.location_id
+left join {{ ref('int_avg_service_rating') }} avg_rtg on main.id = avg_rtg.location_id::int
